+++ miden-crypto/src/rand/mod.rs
@@ -5,11 +5,8 @@ use rand::RngCore;
 
 use crate::{Felt, Word};
 
-mod rpo;
-pub use rpo::RpoRandomCoin;
-
-mod rpx;
-pub use rpx::RpxRandomCoin;
+mod coin;
+pub use coin::RandomCoin;
 
 // Test utilities for generating random data (used in tests and benchmarks)
 #[cfg(any(test, feature = "std"))]